refactor: extract view logic and add mode selector
