+++ app/components/discovery-panel/location-detail.tsx
@@ -1,0 +1,3 @@
+export default function LocationDetail() {
+  return <div>LocationDetail</div>;
+}
