+++ app/components/discovery-panel/navigation-view.tsx
@@ -1,0 +1,3 @@
+export default function NavigationView() {
+  return <div>NavigationView</div>;
+}
