+++ app/types/poi.ts
@@ -1,0 +1,5 @@
+export interface POI {
+  name: string;
+  coordinates: number[];
+  // add any additional properties as needed
+}
